Add support for release package uploading to github for linux/macos
Most of this was generated by travis setup releases, the key is bound to
my account.

For now release-packages are built only for my private repo to avoid
polluting the public one with tags while I play with this.

In order to make sure that the release packages are built with the Release
type I decided to do a second build as part of the before_deploy script.
This also needs the BUNDLE_QT flag for linux/macos packaging of Qt.