Avoid building twice for packages
Since we'll build packages only from tags lets make the normal build a
release-build in that case. Its easy to detect and avoids doing the same
build twice for no particular reason.

In addition to the release/bundling I also disabled setting of
install-rpath as that would leak paths from the system where the build
happens into the packages.