Move the packaging vs. ci invocation of cmake into a script
The logic is becoming to complicated for inlining it into the travis.yml
file, so lets move these bits into a shell script. I don't like having
such a 'magic' script that does the right thing based on the environment
but I don't see a better way right now.

Hopefully the filename makes it clear this is related to detecting some
things in the travis environment.